refactor: request 패키지 분리 및 관련 클래스 구조 변경
- request 관련 정보 추가 및 response 관리를 위한 구조 변경 진행
