+++ src/main/java/webserver/request/RequestUtil.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
